Fix feedback from 2021 update (#3)
This is a follow-up to all the comments I left on #1. Everything still compiles and the tests still pass (i.e., running `make test` in the Examples/ directory shows that the only diffs are in the current date and how long it takes to run).

The [comment about adding back in the statistical checks](https://github.com/penguinland/dealer/pull/1#discussion_r2644824034) was a bigger deal, so is not included here.

The [comment about a possible typo](https://github.com/penguinland/dealer/pull/1#discussion_r2644928342) was so confusing I left it as-is. The sentence makes grammatical sense if "thi" is a noun I don't recognize, but also makes grammatical sense if I replace it with "this many" but not with "this" or "the" or other more obvious typos.



* remove all ._* files

* delete .cvsignore files

* delete binaries and tarfiles of backups

* put the Makefile back to optimized builds and cleaning

* add explanatory comments, tweak whitespace

* use ${PROGRAM} to refer to a program, not a directory

* remove trailing whitespace. There's more in the repo, though

* clean up mystery ^L characters in __random.c

* clean up trailing whitespace in __random.c

* replace leading tabs with 8 spaces in __random.c. I left alone tabs in other formatty places

* fix typo

* delete unused macro

* add comments t o__random.c

* update the MSDOS makefile similar to the normal makefile

* remove trailing whitespace from MSDos makefile, too

* tweak