cast EOF to char before comparing when getting CLI options (#5)
Without this change, it would get into an infinite loop in the CLI options because `c` was 255 instead of -1. I don't see why these values are considered the same on my main computer but not my pi (if you're doing char comparison, 255 and -1 have identical bits, and if you're doing int comparison, a signed char should be extended from 255 to -1. Maybe on a pi chars are unsigned by default? but why isn't that the same on all compilers?), but in practice this was needed.